refactor: Move flip transform utilities to separate flip-utils.ts module
- Create new flip-utils.ts with all flip detection and application logic
- Move buildImageOpIndex, detectFlipFromCtm, flipImageBuffer* functions
- Move applyFlipsToRasterImages function and ImageFlipTransform interface
- Copy hashBuffer to flip-utils to avoid circular dependency
- Update imports in extract.ts to use flip-utils module
- Reduces extract.ts file size and improves code organization
