+++ packages/pdf/src/extract.ts
@@ -6,7 +6,6 @@
 
 import { createHash } from "crypto";
 import { PNG } from "pngjs";
-import jpeg from "jpeg-js";
 import mupdf, {
   type Document as MupdfDocument,
   type PDFDocument,
@@ -38,6 +37,14 @@ import {
 } from "./page-stream-recorder.js";
 import type { DrawItem, PositionedTextOutput } from "@adt/types";
 import { classifyFontCategoryByName } from "@adt/types";
+import {
+  buildImageOpIndex,
+  detectFlipFromCtm,
+  flipImageBufferHorizontal,
+  flipImageBufferVertical,
+  applyFlipsToRasterImages,
+  type ImageFlipTransform,
+} from "./flip-utils.js";
 
 // ============================================================================
 // Types
@@ -561,182 +568,6 @@ function classifyImageColorSpace(resolved: PDFObject): "displayable" | "needs-co
  * Build index of stream ops by native image dimensions.
  * Maps "widthxheight" to ImageStreamOp for quick lookup during extraction.
  */
-function buildImageOpIndex(
-  ops: StreamOp[]
-): Map<string, ImageStreamOp> {
-  const index = new Map<string, ImageStreamOp>();
-  for (const op of ops) {
-    if (op.kind !== "image" && op.kind !== "imageMask") continue;
-    const key = `${op.nativeWidth}x${op.nativeHeight}`;
-    // Store first op with this dimension; further refinement with contentDigest
-    // would happen in stampRasterPlacementsFromOps if needed
-    if (!index.has(key)) {
-      index.set(key, op as ImageStreamOp);
-    }
-  }
-  return index;
-}
-
-/**
- * Detect if image needs horizontal/vertical flip based on CTM.
- * Negative X scale = horizontal flip; negative Y scale = vertical flip.
- */
-interface ImageFlipTransform {
-  flipHorizontal: boolean;
-  flipVertical: boolean;
-}
-
-function detectFlipFromCtm(ctm: number[]): ImageFlipTransform {
-  const [a, , , d] = ctm; // [a, b, c, d, e, f]
-  return {
-    flipHorizontal: a < 0, // Negative X scale
-    flipVertical: d < 0,   // Negative Y scale
-  };
-}
-
-/**
- * Flip image buffer horizontally (left-right mirror).
- * Handles both JPEG and PNG formats.
- */
-function flipImageBufferHorizontal(
-  buffer: Buffer,
-  format: string
-): Buffer {
-  if (format === "jpeg") {
-    const decoded = jpeg.decode(buffer, { useTArray: true });
-    const { width, height } = decoded;
-    const flipped = Buffer.alloc(decoded.data.length);
-
-    // For each row, reverse pixel order (4 bytes per pixel = RGBA)
-    for (let y = 0; y < height; y++) {
-      for (let x = 0; x < width; x++) {
-        const srcIdx = (y * width + x) * 4;
-        const dstIdx = (y * width + (width - 1 - x)) * 4;
-        flipped[dstIdx] = decoded.data[srcIdx];
-        flipped[dstIdx + 1] = decoded.data[srcIdx + 1];
-        flipped[dstIdx + 2] = decoded.data[srcIdx + 2];
-        flipped[dstIdx + 3] = decoded.data[srcIdx + 3];
-      }
-    }
-
-    const encoded = jpeg.encode(
-      { data: flipped, width, height },
-      90 // Maintain quality
-    );
-    return Buffer.from(encoded.data);
-  }
-
-  if (format === "png") {
-    const { data, width, height } = decodePng(buffer);
-    const flipped = Buffer.alloc(data.length);
-
-    // For each row, reverse pixel order (4 bytes per pixel = RGBA)
-    for (let y = 0; y < height; y++) {
-      for (let x = 0; x < width; x++) {
-        const srcIdx = (y * width + x) * 4;
-        const dstIdx = (y * width + (width - 1 - x)) * 4;
-        flipped[dstIdx] = data[srcIdx];
-        flipped[dstIdx + 1] = data[srcIdx + 1];
-        flipped[dstIdx + 2] = data[srcIdx + 2];
-        flipped[dstIdx + 3] = data[srcIdx + 3];
-      }
-    }
-
-    const png = new PNG({ width, height });
-    png.data = flipped;
-    return PNG.sync.write(png);
-  }
-
-  return buffer; // Unknown format
-}
-
-/**
- * Flip image buffer vertically (top-bottom mirror).
- */
-function flipImageBufferVertical(
-  buffer: Buffer,
-  format: string
-): Buffer {
-  if (format === "jpeg") {
-    const decoded = jpeg.decode(buffer, { useTArray: true });
-    const { width, height } = decoded;
-    const flipped = Buffer.alloc(decoded.data.length);
-
-    // Reverse row order
-    for (let y = 0; y < height; y++) {
-      const srcY = height - 1 - y;
-      const srcOffset = srcY * width * 4;
-      const dstOffset = y * width * 4;
-      Buffer.from(decoded.data.buffer).copy(
-        flipped,
-        dstOffset,
-        srcOffset,
-        srcOffset + width * 4
-      );
-    }
-
-    const encoded = jpeg.encode(
-      { data: flipped, width, height },
-      90
-    );
-    return Buffer.from(encoded.data);
-  }
-
-  if (format === "png") {
-    const { data, width, height } = decodePng(buffer);
-    const flipped = Buffer.alloc(data.length);
-
-    for (let y = 0; y < height; y++) {
-      const srcY = height - 1 - y;
-      const srcOffset = srcY * width * 4;
-      const dstOffset = y * width * 4;
-      data.copy(flipped, dstOffset, srcOffset, srcOffset + width * 4);
-    }
-
-    const png = new PNG({ width, height });
-    png.data = flipped;
-    return PNG.sync.write(png);
-  }
-
-  return buffer;
-}
-
-/**
- * Apply CTM-based flip transforms to raster images based on stream operations.
- * Updates image buffers and hashes in-place.
- */
-function applyFlipsToRasterImages(
-  images: ExtractedImage[],
-  streamOps: StreamOp[]
-): void {
-  if (streamOps.length === 0 || images.length === 0) return;
-
-  const opIndex = buildImageOpIndex(streamOps);
-
-  for (const image of images) {
-    const dimKey = `${image.width}x${image.height}`;
-    const streamOp = opIndex.get(dimKey);
-
-    if (!streamOp || !streamOp.ctm) continue;
-
-    const { flipHorizontal, flipVertical } = detectFlipFromCtm(streamOp.ctm);
-    if (!flipHorizontal && !flipVertical) continue;
-
-    let flippedBuf = image.buffer;
-
-    if (flipHorizontal) {
-      flippedBuf = flipImageBufferHorizontal(flippedBuf, image.format);
-    }
-    if (flipVertical) {
-      flippedBuf = flipImageBufferVertical(flippedBuf, image.format);
-    }
-
-    // Update the image with the flipped buffer and recalculate hash
-    image.buffer = flippedBuf;
-    image.hash = hashBuffer(flippedBuf);
-  }
-}
-
 /** Read the /Matte array off an SMask dictionary, expressed as 8-bit RGB.
  * Only RGB (3) and Gray (1) source colorspaces are handled; for anything
  * else (e.g. CMYK) we conservatively skip un-matting. */
